rearrange; 'make' runs slow tests
