+++ .github/workflows/ci.yaml
@@ -15,7 +15,8 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        go-version: [1.19.x, 1.20.x, 1.21.x] # when editing, also update steps and job below
+        # when editing this list, also update steps and jobs below
+        go-version: [1.19.x, 1.20.x, 1.21.x]
     steps:
       - name: Checkout Code
         uses: actions/checkout@v4
@@ -34,10 +35,22 @@ jobs:
         # supported version.
         if: matrix.go-version == '1.21.x'
         run: make checkgenerate && make lint
+  conformance:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        # 1.19 is omitted because conformance test runner requires 1.20+
+        go-version: [1.20.x, 1.21.x]
+    steps:
+      - name: Checkout Code
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 1
+      - name: Install Go
+        uses: actions/setup-go@v4
+        with:
+          go-version: ${{ matrix.go-version }}
       - name: Run Conformance Tests
-        # A dependency of conformance tests (for testing http/3) requires
-        # Go 1.20. So we skip this step for Go 1.19.
-        if: matrix.go-version != '1.19.x'
         run: make runconformance
   slowtest:
     runs-on: ubuntu-latest
